Move desktop app files to package directory
