+++ app-shared/src/commonMain/kotlin/org/pointyware/painteddogs/shared/di/AppDependencies.kt
